chore: trim background-sync narration logs, mock pair format, debug button file
Three focused tightenings on the PR's surface area:

- backgroundSync.ts: drop the stepwise narration logger.log calls
  ("Loading stored ship info", "Configuring urbit client",
  "Running contact discovery", etc.). The trackEvents at entry,
  early-exit, completion, and the timing summary already capture
  the analytics-relevant beats; the narration was useful while
  debugging but adds noise to telemetry and console once shipped.

- lanyardApi.ts: drop the explicit phone:ship pair format from
  EXPO_PUBLIC_MOCK_LANYARD_DISCOVERY. The integer-count form
  covers practically all dev iteration; the pair form was a
  rarely-used affordance for deterministic name testing.

- Extract the dev-only DebugBackgroundSyncButton from App.main.tsx
  into its own component file. Keeps it as a tool for general
  bg-sync debugging (env-gated behind __DEV__ +
  EXPO_PUBLIC_DEBUG_BG_SYNC_BUTTON=true) without inlining ~70
  lines of dev-only React in the app shell.
